Some more restructuring for Python 3
