+++ tests/subtitles.py
@@ -81,7 +81,7 @@ class SrtScriptTestCase(ScriptTestBase):
         self.assertEqual(SINGLE_LINE_SRT_EVENT + "\n\n" + MULTILINE_SRT_EVENT, text)
 
     def test_read_from_file(self):
-        os.write(self.script_description, """1
+        os.write(self.script_description, b"""1
 00:00:17,500 --> 00:00:18,870
 Yeah, really!
 
@@ -131,7 +131,7 @@ Format: Layer, Start, End, Style, Name, MarginL, MarginR, MarginV, Effect, Text
 {0}""".format(ASS_EVENT), text)
 
     def test_read_from_file(self):
-        text = """[Script Info]
+        text = b"""[Script Info]
 ; Script generated by Aegisub 3.1.1
 Title: script title
